# Backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: BeeBa API
  version: 0.1.0
  description: Versioned API for the BeeBa Basis / VR content catalog.
servers:
  - url: http://localhost:8080
    description: Local backend
  - url: https://beeba.org
    description: Production domain
paths:
...
  /api/v1/me/content/{content_id}/images:
    get:
      summary: List preview and gallery images for owned content
      operationId: listOwnedContentImages
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: content_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Owned content images.
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/UploadedImage"
        default:
          $ref: "#/components/responses/Error"
    post:
      summary: Upload a preview or gallery image for owned content
      operationId: uploadOwnedContentImage
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: content_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              allOf:
                - $ref: "#/components/schemas/ImageUploadRequest"
                - type: object
                  properties:
                    is_primary:
                      type: boolean
                      description: When true, this image becomes the content card/detail preview.
                    sort_order:
                      type: integer
                      minimum: 0
                      maximum: 1000
      responses:
        "201":
          description: Content image stored and image processing job queued.
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: "#/components/schemas/UploadedImage"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/me/content/{content_id}/images/{image_id}:
    patch:
      summary: Update metadata for an owned content image
      operationId: updateOwnedContentImage
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: content_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: image_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContentImageUpdateRequest"
      responses:
        "200":
          description: Updated owned content image.
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: "#/components/schemas/UploadedImage"
        default:
          $ref: "#/components/responses/Error"
    delete:
      summary: Remove an owned content image from public and owner galleries
      operationId: deleteOwnedContentImage
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: content_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: image_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Image removed.
        default:
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    cookieAuth:
      type: apiKey
      in: cookie
      name: beeba_access_token
      description: Browser session cookie. Unsafe methods authenticated by cookie also require `X-CSRF-Token` matching the `beeba_csrf_token` cookie.
  responses:
    Error:
      description: Consistent API error response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
            message:
              type: string
            request_id:
              type: string
...
    ImageUploadRequest:
      type: object
      required:
        - file
      properties:
        alt_text:
          type: string
          maxLength: 500
        file:
          type: string
          format: binary
          description: JPEG or PNG image. Stored in preview storage and asynchronously converted to compressed WebP before public serving. The default maximum upload size is 8 MiB unless configured differently.
    UploadedImage:
      type: object
      required:
        - id
        - kind
        - url
        - alt_text
        - width
        - height
        - file_size
        - file_hash_sha256
        - mime_type_detected
        - processing_status
        - created_at
      properties:
        id:
          type: string
          format: uuid
        kind:
          type: string
          enum:
            - user_avatar
            - content_image
        content_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        url:
          type: string
          format: uri-reference
        alt_text:
          type: string
        width:
          type: integer
          minimum: 1
        height:
          type: integer
          minimum: 1
        file_size:
          type: integer
          format: int64
          minimum: 1
        file_hash_sha256:
          type: string
          pattern: "^[a-f0-9]{64}$"
        mime_type_detected:
          type: string
          enum:
            - image/jpeg
            - image/png
            - image/webp
        processing_status:
          type: string
          enum:
            - pending
            - running
            - processed
            - failed
        is_primary:
          type: boolean
        sort_order:
          type: integer
        created_at:
          type: string
          format: date-time
    ContentImageUpdateRequest:
      type: object
      additionalProperties: false
      properties:
        alt_text:
          type: string
          maxLength: 500
        is_primary:
          type: boolean
          description: When true, this image becomes the single primary preview for the content item.
        sort_order:
          type: integer
          minimum: 0
          maximum: 1000
...